pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["cillam <cillam@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
joblib = "1.4.2"
scikit-learn = "1.5.2"
fastapi-cache2 = {extras = ["redis"], version = "0.1.9"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
